LLM/requirement.txt
# Python版本采用Python 3.10.18，运行该项目需要安装的安装包
adalflow==1.1.2
httpx==0.28.1
python-dotenv==1.1.1
mcp==1.13.0
fast_mcp = 2.11.3
openai==1.99.9
asyncio==4.0.0